+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-## v0.34.3 (WIP)
+## v0.35 (WIP)
+
+- Added `nullString()`, `nullInt()`, `nullFloat()`, `nullBool`, `nullArray()`, `nullObject()` JSVM helpers for scanning nullable columns ([#7396](https://github.com/pocketbase/pocketbase/issues/7396)).
+    (@todo update the landing docs)
 
 - Store the correct `image/png` as attrs content type when generating a thumb fallback _(e.g. for `webp`)_.
 
-- Trimmed the normalized file extension to ensure that there is only one beginning `.` character.
+- Trimmed the normalized file extension from leftover `.` characters after cleanup.
 
 - Updated Go deps.